+++ data-raw/schemes_OkabeIto.R
@@ -1,0 +1,15 @@
+# Okabe and Ito color types ====================================================
+schemes_okabe2008 <- list(
+  # Qualitative colour types -------------------------------------------------
+  # The colours have to be used as given: do not interpolate
+  "okabe ito" = list(
+    type = "qualitative",
+    interpolate = FALSE,
+    missing = NA_character_,
+    colours = c(
+      `black`        = "#000000", `orange` = "#E69F00", `sky blue`  = "#56B4E9",
+      `bluish green` = "#009E73", `yellow` = "#F0E442", `blue`      = "#0072B2",
+      `vermilion`    = "#D55E00", `reddish purple` = "#CC79A7"),
+    scheme = NULL
+  )
+)