+++ inst/examples/ex-okabeito-discrete.R
@@ -1,0 +1,5 @@
+library(ggplot2)
+
+ggplot2::ggplot(mpg, ggplot2::aes(displ, hwy, colour = class)) +
+  ggplot2::geom_point() +
+  scale_colour_okabeito()