+++ tests/testthat/test-ggplot2-okabeito.R
@@ -1,0 +1,22 @@
+context("ggplot2 Okabe and Ito scale")
+
+# Qualitative color schemes ====================================================
+test_that("Bright", {
+  skip_if_not_installed("ggplot2")
+  expect_s3_class(scale_colour_okabeito(), "ggproto")
+  expect_s3_class(scale_colour_okabeito(), "ScaleDiscrete")
+  expect_equal(
+    scale_colour_okabeito()$palette(8),
+    colour("okabe ito", names = FALSE)(8)
+  )
+
+  expect_s3_class(scale_color_okabeito(), "ggproto")
+  expect_s3_class(scale_color_okabeito(), "ScaleDiscrete")
+
+  expect_s3_class(scale_fill_okabeito(), "ggproto")
+  expect_s3_class(scale_fill_okabeito(), "ScaleDiscrete")
+  expect_equal(
+    scale_fill_okabeito()$palette(8),
+    colour("okabe ito", names = FALSE)(8)
+  )
+})